refactor: 블랙잭 게임 클래스 추가 및 프로젝트 구조 변경
